+++ rclcpp_action/include/rclcpp_action/create_client.hpp
@@ -30,13 +30,14 @@ namespace rclcpp_action
  * This function is equivalent to \sa create_client()` however is using the individual
  * node interfaces to create the client.
  *
- * \param node_base_interface[in] The node base interface of the corresponding node.
- * \param node_graph_interface[in] The node graph interface of the corresponding node.
- * \param node_logging_interface[in] The node logging interface of the corresponding node.
- * \param node_waitables_interface[in] The node waitables interface of the corresponding node.
+ * \param[in] node_base_interface The node base interface of the corresponding node.
+ * \param[in] node_graph_interface The node graph interface of the corresponding node.
+ * \param[in] node_logging_interface The node logging interface of the corresponding node.
+ * \param[in] node_waitables_interface The node waitables interface of the corresponding node.
  * \param[in] name The action name.
  * \param[in] group The action client will be added to this callback group.
  *   If `nullptr`, then the action client is added to the default callback group.
+ * \param[in] options Options to pass to the underlying `rcl_action_client_t`.
  */
 template<typename ActionT>
 typename Client<ActionT>::SharedPtr
@@ -46,7 +47,8 @@ create_client(
   rclcpp::node_interfaces::NodeLoggingInterface::SharedPtr node_logging_interface,
   rclcpp::node_interfaces::NodeWaitablesInterface::SharedPtr node_waitables_interface,
   const std::string & name,
-  rclcpp::CallbackGroup::SharedPtr group = nullptr)
+  rclcpp::CallbackGroup::SharedPtr group = nullptr,
+  const rcl_action_client_options_t & options = rcl_action_client_get_default_options())
 {
   std::weak_ptr<rclcpp::node_interfaces::NodeWaitablesInterface> weak_node =
     node_waitables_interface;
@@ -59,20 +61,19 @@ create_client(
         return;
       }
       auto shared_node = weak_node.lock();
-      if (!shared_node) {
-        return;
-      }
-      // API expects a shared pointer, give it one with a deleter that does nothing.
-      std::shared_ptr<Client<ActionT>> fake_shared_ptr(ptr, [](Client<ActionT> *) {});
+      if (shared_node) {
+        // API expects a shared pointer, give it one with a deleter that does nothing.
+        std::shared_ptr<Client<ActionT>> fake_shared_ptr(ptr, [](Client<ActionT> *) {});
 
-      if (group_is_null) {
-        // Was added to default group
-        shared_node->remove_waitable(fake_shared_ptr, nullptr);
-      } else {
-        // Was added to a specfic group
-        auto shared_group = weak_group.lock();
-        if (shared_group) {
-          shared_node->remove_waitable(fake_shared_ptr, shared_group);
+        if (group_is_null) {
+          // Was added to default group
+          shared_node->remove_waitable(fake_shared_ptr, nullptr);
+        } else {
+          // Was added to a specific group
+          auto shared_group = weak_group.lock();
+          if (shared_group) {
+            shared_node->remove_waitable(fake_shared_ptr, shared_group);
+          }
         }
       }
       delete ptr;
@@ -83,7 +84,8 @@ create_client(
       node_base_interface,
       node_graph_interface,
       node_logging_interface,
-      name),
+      name,
+      options),
     deleter);
 
   node_waitables_interface->add_waitable(action_client, group);
@@ -96,13 +98,15 @@ create_client(
  * \param[in] name The action name.
  * \param[in] group The action client will be added to this callback group.
  *   If `nullptr`, then the action client is added to the default callback group.
+ * \param[in] options Options to pass to the underlying `rcl_action_client_t`.
  */
 template<typename ActionT, typename NodeT>
 typename Client<ActionT>::SharedPtr
 create_client(
   NodeT node,
   const std::string & name,
-  rclcpp::CallbackGroup::SharedPtr group = nullptr)
+  rclcpp::CallbackGroup::SharedPtr group = nullptr,
+  const rcl_action_client_options_t & options = rcl_action_client_get_default_options())
 {
   return rclcpp_action::create_client<ActionT>(
     node->get_node_base_interface(),
@@ -110,7 +114,8 @@ create_client(
     node->get_node_logging_interface(),
     node->get_node_waitables_interface(),
     name,
-    group);
+    group,
+    options);
 }
 }  // namespace rclcpp_action